Added "required" flag
We have existing software running that does not support JWT. We need the ability to add support for JWT to new versions of our API and software and not break existing clients. The "required" flag is used to indicate that the token should be processed if it's present, but should not cause the request to fail if it is not.